+++ packages/chrome/src/theme.js
@@ -24,7 +24,8 @@ export default {
   },
   fontSizes: [12, 14, 16, 20, 24, 32, 48, 64, 96, 128],
   fonts: {
-    body: 'system-ui, sans-serif',
+    body:
+      'system-ui, -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, "Helvetica Neue", sans-serif',
     heading: 'inherit',
     monospace: 'Menlo, monospace',
   },