+++ packages/editor/src/Theme/Fonts.js
@@ -3,7 +3,7 @@ import { jsx, useThemeUI } from 'theme-ui'
 import Combobox from '../Combobox'
 
 const defaultFonts = [
-  'system-ui, sans-serif',
+  'system-ui, -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, "Helvetica Neue", sans-serif',
   '"Helvetica Neue", Helvetica, Arial, sans-serif',
   '"Avenir Next", Helvetica, Arial, sans-serif',
   'Verdana, sans-serif',